+++ tests/parser/features/test_internal_call.py
@@ -2,8 +2,14 @@ from decimal import (
     Decimal,
 )
 
+import pytest
+
+from vyper import (
+    compiler,
+)
 from vyper.exceptions import (
     StructureException,
+    TypeMismatchException,
 )
 
 
@@ -36,7 +42,7 @@ def returnten() -> int128:
 
 @public
 def _hashy(x: bytes32) -> bytes32:
-    return sha3(x)
+    return keccak256(x)
 
 @public
 def return_hash_of_rzpadded_cow() -> bytes32:
@@ -54,7 +60,7 @@ def test_selfcall_code_3(get_contract_with_gas_estimation, keccak):
     selfcall_code_3 = """
 @public
 def _hashy2(x: bytes[100]) -> bytes32:
-    return sha3(x)
+    return keccak256(x)
 
 @public
 def return_hash_of_cow_x_30() -> bytes32:
@@ -324,6 +330,30 @@ def bar() -> (int128, decimal):
     assert c.bar() == [66, Decimal('66.77')]
 
 
+def test_private_function_multiple_lists_as_args(get_contract_with_gas_estimation):
+    code = """
+@private
+def _foo(y: int128[2], x: bytes[5]) -> int128:
+    return y[0]
+
+@private
+def _foo2(x: bytes[5], y: int128[2]) -> int128:
+    return y[0]
+
+@public
+def bar() -> int128:
+    return self._foo([1, 2], b"hello")
+
+@public
+def bar2() -> int128:
+    return self._foo2(b"hello", [1, 2])
+"""
+
+    c = get_contract_with_gas_estimation(code)
+    assert c.bar() == 1
+    assert c.bar2() == 1
+
+
 def test_multi_mixed_arg_list_bytes_call(get_contract_with_gas_estimation):
     code = """
 @public
@@ -362,3 +392,55 @@ def foo() -> int128:
     return self.bar(1)
 """
     assert_tx_failed(lambda: get_contract_with_gas_estimation(code), StructureException)
+
+
+FAILING_CONTRACTS = [
+    """
+# should not compile - value kwarg when calling {0} function
+@{0}
+def foo():
+    pass
+
+@public
+def bar():
+    self.foo(value=100)
+    """,
+    """
+# should not compile - gas kwarg when calling {0} function
+@{0}
+def foo():
+    pass
+
+@public
+def bar():
+    self.foo(gas=100)
+    """,
+    """
+# should not compile - arbitrary kwargs when calling {0} function
+@{0}
+def foo():
+    pass
+
+@public
+def bar():
+    self.foo(baz=100)
+    """,
+    """
+# should not compile - args-as-kwargs to a {0} function
+@{0}
+def foo(baz: int128):
+    pass
+
+@public
+def bar():
+    self.foo(baz=100)
+    """,
+]
+
+
+@pytest.mark.parametrize('failing_contract_code', FAILING_CONTRACTS)
+@pytest.mark.parametrize('decorator', ['public', 'private'])
+def test_selfcall_kwarg_raises(failing_contract_code, decorator):
+    code = failing_contract_code.format(decorator)
+    with pytest.raises(TypeMismatchException):
+        compiler.compile_code(code)