+++ tests/parser/integration/test_basics.py
@@ -25,7 +25,7 @@ def test_selfcall_code_3(get_contract_with_gas_estimation, keccak):
     selfcall_code_3 = """
 @public
 def _hashy2(x: bytes[100]) -> bytes32:
-    return sha3(x)
+    return keccak256(x)
 
 @public
 def return_hash_of_cow_x_30() -> bytes32: