+++ vyper/parser/self_call.py
@@ -2,6 +2,7 @@ import itertools
 
 from vyper.exceptions import (
     ConstancyViolationException,
+    TypeMismatchException,
 )
 from vyper.parser.lll_node import (
     LLLnode,
@@ -26,9 +27,27 @@ from vyper.types import (
 
 def call_lookup_specs(stmt_expr, context):
     from vyper.parser.expr import Expr
+
     method_name = stmt_expr.func.attr
-    expr_args = [Expr(arg, context).lll_node for arg in stmt_expr.args]
-    sig = FunctionSignature.lookup_sig(context.sigs, method_name, expr_args, stmt_expr, context)
+
+    if len(stmt_expr.keywords):
+        raise TypeMismatchException(
+            "Cannot use keyword arguments in calls to functions via 'self'",
+            stmt_expr,
+        )
+    expr_args = [
+        Expr(arg, context).lll_node
+        for arg in stmt_expr.args
+    ]
+
+    sig = FunctionSignature.lookup_sig(
+        context.sigs,
+        method_name,
+        expr_args,
+        stmt_expr,
+        context,
+    )
+
     return method_name, expr_args, sig
 
 
@@ -82,16 +101,39 @@ def call_self_private(stmt_expr, context, sig):
     push_args = []
 
     # Push local variables.
-    if context.vars:
-        var_slots = [(v.pos, v.size) for name, v in context.vars.items()]
+    var_slots = [
+        (v.pos, v.size) for name, v in context.vars.items()
+        if v.location == 'memory'
+    ]
+    if var_slots:
         var_slots.sort(key=lambda x: x[0])
         mem_from, mem_to = var_slots[0][0], var_slots[-1][0] + var_slots[-1][1] * 32
-        push_local_vars = [
-            ['mload', pos] for pos in range(mem_from, mem_to, 32)
-        ]
-        pop_local_vars = [
-            ['mstore', pos, 'pass'] for pos in reversed(range(mem_from, mem_to, 32))
-        ]
+
+        i_placeholder = context.new_placeholder(BaseType('uint256'))
+        local_save_ident = "_%d_%d" % (stmt_expr.lineno, stmt_expr.col_offset)
+        push_loop_label = 'save_locals_start' + local_save_ident
+        pop_loop_label = 'restore_locals_start' + local_save_ident
+
+        if mem_to - mem_from > 320:
+            push_local_vars = [
+                    ['mstore', i_placeholder, mem_from],
+                    ['label', push_loop_label],
+                    ['mload', ['mload', i_placeholder]],
+                    ['mstore', i_placeholder, ['add', ['mload', i_placeholder], 32]],
+                    ['if', ['lt', ['mload', i_placeholder], mem_to],
+                        ['goto', push_loop_label]]
+            ]
+            pop_local_vars = [
+                ['mstore', i_placeholder, mem_to - 32],
+                ['label', pop_loop_label],
+                ['mstore', ['mload', i_placeholder], 'pass'],
+                ['mstore', i_placeholder, ['sub', ['mload', i_placeholder], 32]],
+                ['if', ['ge', ['mload', i_placeholder], mem_from],
+                       ['goto', pop_loop_label]]
+            ]
+        else:
+            push_local_vars = [['mload', pos] for pos in range(mem_from, mem_to, 32)]
+            pop_local_vars = [['mstore', pos, 'pass'] for pos in range(mem_to-32, mem_from-32, -32)]
 
     # Push Arguments
     if expr_args:
@@ -107,7 +149,7 @@ def call_self_private(stmt_expr, context, sig):
         static_arg_size = 32 * sum(
                 [get_static_size_of_type(arg.typ)
                     for arg in expr_args])
-        static_pos = arg_pos + static_arg_size
+        static_pos = int(arg_pos + static_arg_size)
         needs_dyn_section = any(
                 [has_dynamic_data(arg.typ)
                     for arg in expr_args])
@@ -126,9 +168,11 @@ def call_self_private(stmt_expr, context, sig):
             # by taking ceil32(len<arg>) + offset<arg> + arg_pos
             # for the last dynamic argument and arg_pos is the start
             # the whole argument section.
-            for idx, arg in enumerate(expr_args):
+            idx = 0
+            for arg in expr_args:
                 if isinstance(arg.typ, ByteArrayLike):
                     last_idx = idx
+                idx += get_static_size_of_type(arg.typ)
             push_args += [
                 ['with', 'offset', ['mload', arg_pos + last_idx * 32],
                     ['with', 'len_pos', ['add', arg_pos, 'offset'],